renew

+++ fe/fe-core/src/main/java/org/apache/doris/datasource/hive/HMSExternalCatalog.java
@@ -17,34 +17,52 @@
 
 package org.apache.doris.datasource.hive;
 
+import org.apache.doris.analysis.TableValuedFunctionRef;
 import org.apache.doris.catalog.Env;
 import org.apache.doris.catalog.HdfsResource;
 import org.apache.doris.cluster.ClusterNamespace;
 import org.apache.doris.common.Config;
 import org.apache.doris.common.DdlException;
+import org.apache.doris.common.Pair;
+import org.apache.doris.common.ThreadPoolManager;
 import org.apache.doris.common.security.authentication.AuthenticationConfig;
-import org.apache.doris.common.security.authentication.HadoopUGI;
+import org.apache.doris.common.security.authentication.HadoopAuthenticator;
+import org.apache.doris.common.util.Util;
 import org.apache.doris.datasource.CatalogProperty;
 import org.apache.doris.datasource.ExternalCatalog;
 import org.apache.doris.datasource.ExternalDatabase;
 import org.apache.doris.datasource.ExternalTable;
 import org.apache.doris.datasource.InitCatalogLog;
 import org.apache.doris.datasource.SessionContext;
+import org.apache.doris.datasource.iceberg.IcebergUtils;
 import org.apache.doris.datasource.jdbc.client.JdbcClientConfig;
 import org.apache.doris.datasource.operations.ExternalMetadataOperations;
 import org.apache.doris.datasource.property.PropertyConverter;
 import org.apache.doris.datasource.property.constants.HMSProperties;
+import org.apache.doris.fs.FileSystemProvider;
+import org.apache.doris.fs.FileSystemProviderImpl;
+import org.apache.doris.fs.remote.dfs.DFSFileSystem;
+import org.apache.doris.nereids.exceptions.AnalysisException;
+import org.apache.doris.nereids.trees.expressions.functions.table.PartitionValues;
+import org.apache.doris.nereids.trees.expressions.functions.table.TableValuedFunction;
+import org.apache.doris.transaction.TransactionManagerFactory;
 
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
+import lombok.Getter;
 import org.apache.commons.lang3.math.NumberUtils;
 import org.apache.hadoop.hive.conf.HiveConf;
+import org.apache.iceberg.hive.HiveCatalog;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
+import java.util.concurrent.ThreadPoolExecutor;
 
 /**
  * External catalog for hive metastore compatible data sources.
@@ -55,13 +73,24 @@ public class HMSExternalCatalog extends ExternalCatalog {
     public static final String FILE_META_CACHE_TTL_SECOND = "file.meta.cache.ttl-second";
     // broker name for file split and query scan.
     public static final String BIND_BROKER_NAME = "broker.name";
-    private static final String PROP_ALLOW_FALLBACK_TO_SIMPLE_AUTH = "ipc.client.fallback-to-simple-auth-allowed";
 
     // -1 means file cache no ttl set
     public static final int FILE_META_CACHE_NO_TTL = -1;
     // 0 means file cache is disabled; >0 means file cache with ttl;
     public static final int FILE_META_CACHE_TTL_DISABLE_CACHE = 0;
 
+    private static final int FILE_SYSTEM_EXECUTOR_THREAD_NUM = 16;
+    private ThreadPoolExecutor fileSystemExecutor;
+    @Getter
+    private HadoopAuthenticator authenticator;
+
+    private int hmsEventsBatchSizePerRpc = -1;
+    private boolean enableHmsEventsIncrementalSync = false;
+
+    //for "type" = "hms" , but is iceberg table.
+    private HiveCatalog icebergHiveCatalog;
+
+    @VisibleForTesting
     public HMSExternalCatalog() {
         catalogProperty = new CatalogProperty(null, null);
     }
@@ -74,6 +103,8 @@ public class HMSExternalCatalog extends ExternalCatalog {
         super(catalogId, name, InitCatalogLog.Type.HMS, comment);
         props = PropertyConverter.convertToMetaProperties(props);
         catalogProperty = new CatalogProperty(resource, props);
+        AuthenticationConfig config = AuthenticationConfig.getKerberosConfig(getConfiguration());
+        authenticator = HadoopAuthenticator.getHadoopAuthenticator(config);
     }
 
     @Override
@@ -85,6 +116,19 @@ public class HMSExternalCatalog extends ExternalCatalog {
                 < FILE_META_CACHE_TTL_DISABLE_CACHE) {
             throw new DdlException(
                     "The parameter " + FILE_META_CACHE_TTL_SECOND + " is wrong, value is " + fileMetaCacheTtlSecond);
+        }
+        Map<String, String> properties = catalogProperty.getProperties();
+        if (properties.containsKey(HMSProperties.ENABLE_HMS_EVENTS_INCREMENTAL_SYNC)) {
+            enableHmsEventsIncrementalSync =
+                    properties.get(HMSProperties.ENABLE_HMS_EVENTS_INCREMENTAL_SYNC).equals("true");
+        } else {
+            enableHmsEventsIncrementalSync = Config.enable_hms_events_incremental_sync;
+        }
+
+        if (properties.containsKey(HMSProperties.HMS_EVENTIS_BATCH_SIZE_PER_RPC)) {
+            hmsEventsBatchSizePerRpc = Integer.valueOf(properties.get(HMSProperties.HMS_EVENTIS_BATCH_SIZE_PER_RPC));
+        } else {
+            hmsEventsBatchSizePerRpc = Config.hms_events_batch_size_per_rpc;
         }
 
         // check the dfs.ha properties
@@ -124,6 +168,11 @@ public class HMSExternalCatalog extends ExternalCatalog {
 
     @Override
     protected void initLocalObjectsImpl() {
+        if (authenticator == null) {
+            AuthenticationConfig config = AuthenticationConfig.getKerberosConfig(getConfiguration());
+            authenticator = HadoopAuthenticator.getHadoopAuthenticator(config);
+        }
+
         HiveConf hiveConf = null;
         JdbcClientConfig jdbcClientConfig = null;
         String hiveMetastoreType = catalogProperty.getOrDefault(HMSProperties.HIVE_METASTORE_TYPE, "");
@@ -141,24 +190,31 @@ public class HMSExternalCatalog extends ExternalCatalog {
             }
             hiveConf.set(HiveConf.ConfVars.METASTORE_CLIENT_SOCKET_TIMEOUT.name(),
                     String.valueOf(Config.hive_metastore_client_timeout_second));
-            HadoopUGI.tryKrbLogin(this.getName(), AuthenticationConfig.getKerberosConfig(hiveConf,
-                    AuthenticationConfig.HIVE_KERBEROS_PRINCIPAL,
-                    AuthenticationConfig.HIVE_KERBEROS_KEYTAB));
         }
-        metadataOps = ExternalMetadataOperations.newHiveMetadataOps(hiveConf, jdbcClientConfig, this);
+        HiveMetadataOps hiveOps = ExternalMetadataOperations.newHiveMetadataOps(hiveConf, jdbcClientConfig, this);
+        FileSystemProvider fileSystemProvider = new FileSystemProviderImpl(Env.getCurrentEnv().getExtMetaCacheMgr(),
+                this.bindBrokerName(), this.catalogProperty.getHadoopProperties());
+        this.fileSystemExecutor = ThreadPoolManager.newDaemonFixedThreadPool(FILE_SYSTEM_EXECUTOR_THREAD_NUM,
+                Integer.MAX_VALUE, String.format("hms_committer_%s_file_system_executor_pool", name), true);
+        transactionManager = TransactionManagerFactory.createHiveTransactionManager(hiveOps, fileSystemProvider,
+                fileSystemExecutor);
+        metadataOps = hiveOps;
+
+        icebergHiveCatalog = IcebergUtils.createIcebergHiveCatalog(this, getName());
+    }
+
+    @Override
+    public void onRefresh(boolean invalidCache) {
+        super.onRefresh(invalidCache);
+        if (metadataOps != null) {
+            metadataOps.close();
+        }
     }
 
     @Override
     public List<String> listTableNames(SessionContext ctx, String dbName) {
         makeSureInitialized();
-        HMSExternalDatabase hmsExternalDatabase = (HMSExternalDatabase) idToDb.get(dbNameToId.get(dbName));
-        if (hmsExternalDatabase != null && hmsExternalDatabase.isInitialized()) {
-            List<String> names = Lists.newArrayList();
-            hmsExternalDatabase.getTables().forEach(table -> names.add(table.getName()));
-            return names;
-        } else {
-            return metadataOps.listTableNames(ClusterNamespace.getNameFromFullName(dbName));
-        }
+        return metadataOps.listTableNames(ClusterNamespace.getNameFromFullName(dbName));
     }
 
     @Override
@@ -169,7 +225,7 @@ public class HMSExternalCatalog extends ExternalCatalog {
     @Override
     public boolean tableExistInLocal(String dbName, String tblName) {
         makeSureInitialized();
-        HMSExternalDatabase hmsExternalDatabase = (HMSExternalDatabase) idToDb.get(dbNameToId.get(dbName));
+        HMSExternalDatabase hmsExternalDatabase = (HMSExternalDatabase) getDbNullable(dbName);
         if (hmsExternalDatabase == null) {
             return false;
         }
@@ -186,11 +242,17 @@ public class HMSExternalCatalog extends ExternalCatalog {
         if (LOG.isDebugEnabled()) {
             LOG.debug("drop database [{}]", dbName);
         }
-        Long dbId = dbNameToId.remove(dbName);
-        if (dbId == null) {
-            LOG.warn("drop database [{}] failed", dbName);
+        if (useMetaCache.get()) {
+            if (isInitialized()) {
+                metaCache.invalidate(dbName, Util.genIdByName(getQualifiedName(dbName)));
+            }
+        } else {
+            Long dbId = dbNameToId.remove(dbName);
+            if (dbId == null) {
+                LOG.warn("drop database [{}] failed", dbName);
+            }
+            idToDb.remove(dbId);
         }
-        idToDb.remove(dbId);
         Env.getCurrentEnv().getExtMetaCacheMgr().invalidateDbCache(getId(), dbName);
     }
 
@@ -199,9 +261,16 @@ public class HMSExternalCatalog extends ExternalCatalog {
         if (LOG.isDebugEnabled()) {
             LOG.debug("create database [{}]", dbName);
         }
-        dbNameToId.put(dbName, dbId);
-        ExternalDatabase<? extends ExternalTable> db = getDbForInit(dbName, dbId, logType);
-        idToDb.put(dbId, db);
+
+        ExternalDatabase<? extends ExternalTable> db = buildDbForInit(dbName, dbId, logType, false);
+        if (useMetaCache.get()) {
+            if (isInitialized()) {
+                metaCache.updateCache(dbName, db, Util.genIdByName(getQualifiedName(dbName)));
+            }
+        } else {
+            dbNameToId.put(dbName, dbId);
+            idToDb.put(dbId, db);
+        }
     }
 
     @Override
@@ -214,14 +283,42 @@ public class HMSExternalCatalog extends ExternalCatalog {
     }
 
     @Override
-    public void setDefaultPropsWhenCreating(boolean isReplay) {
-        if (isReplay) {
-            return;
-        }
-        if (catalogProperty.getOrDefault(PROP_ALLOW_FALLBACK_TO_SIMPLE_AUTH, "").isEmpty()) {
+    public void setDefaultPropsIfMissing(boolean isReplay) {
+        super.setDefaultPropsIfMissing(isReplay);
+        if (ifNotSetFallbackToSimpleAuth()) {
             // always allow fallback to simple auth, so to support both kerberos and simple auth
-            catalogProperty.addProperty(PROP_ALLOW_FALLBACK_TO_SIMPLE_AUTH, "true");
+            catalogProperty.addProperty(DFSFileSystem.PROP_ALLOW_FALLBACK_TO_SIMPLE_AUTH, "true");
         }
+    }
+
+    @Override
+    public Pair<String, String> getSourceTableNameWithMetaTableName(String tableName) {
+        for (MetaTableFunction metaFunction : MetaTableFunction.values()) {
+            if (metaFunction.containsMetaTable(tableName)) {
+                return Pair.of(metaFunction.getSourceTableName(tableName), metaFunction.name().toLowerCase());
+            }
+        }
+        return Pair.of(tableName, "");
+    }
+
+    @Override
+    public Optional<TableValuedFunction> getMetaTableFunction(String dbName, String sourceNameWithMetaName) {
+        for (MetaTableFunction metaFunction : MetaTableFunction.values()) {
+            if (metaFunction.containsMetaTable(sourceNameWithMetaName)) {
+                return Optional.of(metaFunction.createFunction(name, dbName, sourceNameWithMetaName));
+            }
+        }
+        return Optional.empty();
+    }
+
+    @Override
+    public Optional<TableValuedFunctionRef> getMetaTableFunctionRef(String dbName, String sourceNameWithMetaName) {
+        for (MetaTableFunction metaFunction : MetaTableFunction.values()) {
+            if (metaFunction.containsMetaTable(sourceNameWithMetaName)) {
+                return Optional.of(metaFunction.createFunctionRef(name, dbName, sourceNameWithMetaName));
+            }
+        }
+        return Optional.empty();
     }
 
     public String getHiveMetastoreUris() {
@@ -231,4 +328,70 @@ public class HMSExternalCatalog extends ExternalCatalog {
     public String getHiveVersion() {
         return catalogProperty.getOrDefault(HMSProperties.HIVE_VERSION, "");
     }
+
+    public int getHmsEventsBatchSizePerRpc() {
+        return hmsEventsBatchSizePerRpc;
+    }
+
+    public boolean isEnableHmsEventsIncrementalSync() {
+        return enableHmsEventsIncrementalSync;
+    }
+
+    public HiveCatalog getIcebergHiveCatalog() {
+        return icebergHiveCatalog;
+    }
+
+    /**
+     * Enum for meta tables in hive catalog.
+     * eg: tbl$partitions
+     */
+    private enum MetaTableFunction {
+        PARTITIONS("partition_values");
+
+        private final String suffix;
+        private final String tvfName;
+
+        MetaTableFunction(String tvfName) {
+            this.suffix = "$" + name().toLowerCase();
+            this.tvfName = tvfName;
+        }
+
+        boolean containsMetaTable(String tableName) {
+            return tableName.endsWith(suffix) && (tableName.length() > suffix.length());
+        }
+
+        String getSourceTableName(String tableName) {
+            return tableName.substring(0, tableName.length() - suffix.length());
+        }
+
+        public TableValuedFunction createFunction(String ctlName, String dbName, String sourceNameWithMetaName) {
+            switch (this) {
+                case PARTITIONS:
+                    List<String> nameParts = Lists.newArrayList(ctlName, dbName,
+                            getSourceTableName(sourceNameWithMetaName));
+                    return PartitionValues.create(nameParts);
+                default:
+                    throw new AnalysisException("Unsupported meta function type: " + this);
+            }
+        }
+
+        public TableValuedFunctionRef createFunctionRef(String ctlName, String dbName, String sourceNameWithMetaName) {
+            switch (this) {
+                case PARTITIONS:
+                    Map<String, String> params = Maps.newHashMap();
+                    params.put("catalog", ctlName);
+                    params.put("database", dbName);
+                    params.put("table", getSourceTableName(sourceNameWithMetaName));
+                    try {
+                        return new TableValuedFunctionRef(tvfName, null, params);
+                    } catch (org.apache.doris.common.AnalysisException e) {
+                        LOG.warn("should not happen. {}.{}.{}", ctlName, dbName, sourceNameWithMetaName);
+                        return null;
+                    }
+                default:
+                    throw new AnalysisException("Unsupported meta function type: " + this);
+            }
+        }
+    }
 }
+
